# swagger.yaml
openapi: 3.0.3
info:
  title: IDM App API
  description: |
    Identity and Access Management (IDM) Application API
    
    This API provides comprehensive user, group, organization, and role management capabilities
    with Auth0 integration for secure authentication and authorization.
    
    ## Features
    - User Management (CRUD operations)
    - Group Management (CRUD operations)
    - Organization Management (CRUD operations)
    - Role Management (CRUD operations)
    - Group Member Management (User-Group relationships)
    - Organization Member Management (User/Group/Role-Organization relationships)
    - Role Member Management (User/Group-Role relationships)
    - Auth0 JWT Authentication
    - Health Monitoring
    
    ## Authentication
    All protected endpoints require a valid JWT token from Auth0. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Getting Started
    1. Configure your Auth0 credentials in the `.env` file
    2. Obtain a JWT token using the Auth0 client credentials flow
    3. Use the token to access protected endpoints
    
    ## Note on Missing Endpoints
    The following endpoints exist in the codebase but are not currently wired up in the routes:
    - Organization Member Management (`/api/v1/orgmembers`) - Add/remove users, groups, or roles from organizations
    - Role Member Management (`/api/v1/rolemembers`) - Add/remove users or groups from roles
    
...
    
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/orgs:
    get:
      tags:
        - Organizations
      summary: Get all organizations
      description: Retrieve a list of all organizations in the system
      operationId: getAllOrgs
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of organizations retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrgResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Organizations
      summary: Create a new organization
      description: Create a new organization in the system
      operationId: createOrg
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgCreateRequest'
      responses:
        '201':
          description: Organization created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgResponse'
        '400':
          description: Bad request - Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Organization with this name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/orgs/{id}:
    get:
      tags:
        - Organizations
      summary: Get organization by ID
      description: Retrieve a specific organization by its ID
      operationId: getOrgById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Organization ID (UUID)
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Organization retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Organizations
      summary: Update organization
      description: Update an existing organization's information
      operationId: updateOrg
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Organization ID (UUID)
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgUpdateRequest'
      responses:
        '200':
          description: Organization updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgResponse'
        '400':
          description: Bad request - Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Organization name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Organizations
      summary: Delete organization
      description: Delete an organization from the system
      operationId: deleteOrg
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Organization ID (UUID)
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Organization deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Organization deleted successfully"
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/roles:
    get:
      tags:
        - Roles
      summary: Get all roles
      description: Retrieve a list of all roles in the system
      operationId: getAllRoles
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of roles retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Roles
      summary: Create a new role
      description: Create a new role in the system
      operationId: createRole
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCreateRequest'
      responses:
        '201':
          description: Role created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          description: Bad request - Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Role with this name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/roles/{id}:
    get:
      tags:
        - Roles
      summary: Get role by ID
      description: Retrieve a specific role by its ID
      operationId: getRoleById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Role ID (UUID)
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Role retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Roles
      summary: Update role
      description: Update an existing role's information
      operationId: updateRole
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Role ID (UUID)
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdateRequest'
      responses:
        '200':
          description: Role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          description: Bad request - Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Role name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Roles
      summary: Delete role
      description: Delete a role from the system
      operationId: deleteRole
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Role ID (UUID)
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Role deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Role deleted successfully"
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Role not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from Auth0. Include the token in the Authorization header.
        Example: `Authorization: Bearer eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...`
  schemas:
...
    OrgCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Unique organization name
          example: "acme-corp"
        displayName:
          type: string
          description: Display name for the organization
          example: "ACME Corporation"
        description:
          type: string
          description: Organization description
          example: "A leading technology company"
    OrgUpdateRequest:
      type: object
      properties:
        name:
          type: string
          description: Unique organization name
          example: "acme-enterprise"
        displayName:
          type: string
          description: Display name for the organization
          example: "ACME Enterprise Solutions"
        description:
          type: string
          description: Organization description
          example: "Enterprise technology solutions provider"
        isActive:
          type: boolean
          description: Whether the organization is active
          example: true
    OrgResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Organization ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          description: Organization name
          example: "acme-corp"
        displayName:
          type: string
          description: Display name
          example: "ACME Corporation"
        description:
          type: string
          description: Organization description
          example: "A leading technology company"
        isActive:
          type: boolean
          description: Active status
          example: true
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2025-06-22T14:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2025-06-22T14:00:00Z"
    RoleCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Unique role name
          example: "admin"
        displayName:
          type: string
          description: Display name for the role
          example: "Administrator"
        description:
          type: string
          description: Role description
          example: "System administrator role"
    RoleUpdateRequest:
      type: object
      properties:
        name:
          type: string
          description: Unique role name
          example: "super-admin"
        displayName:
          type: string
          description: Display name for the role
          example: "Super Administrator"
        description:
          type: string
          description: Role description
          example: "Super administrator role with full access"
        isActive:
          type: boolean
          description: Whether the role is active
          example: true
    RoleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Role ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          description: Role name
          example: "admin"
        displayName:
          type: string
          description: Display name
          example: "Administrator"
        description:
          type: string
          description: Role description
          example: "System administrator role"
        isActive:
          type: boolean
          description: Active status
          example: true
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2025-06-22T14:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2025-06-22T14:00:00Z"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "User not found"
        code:
          type: string
          description: Error code
          example: "USER_NOT_FOUND"
        details:
          type: object
          description: Additional error details
          additionalProperties: true
...